Implemented Access Service Layer
The API can now grab information from the database and display it as JSON objects. I have a function for inserting transactions as well, if we so choose to use it. To run the code, you must first create the database table and stored procedures using the scripts in the folder "SQL Scripts". They are numbered by order of execution.